+++ weblate/utils/tests/test_search.py
@@ -109,10 +109,10 @@ class QueryParserTest(TestCase, SearchMixin):
         self.assert_query("location:hello.c", Q(location__substring="hello.c"))
 
     def test_exact(self):
-        self.assert_query("source:='hello'", Q(source__iexact="hello"))
-        self.assert_query('source:="hello world"', Q(source__iexact="hello world"))
-        self.assert_query("source:='hello world'", Q(source__iexact="hello world"))
-        self.assert_query("source:=hello", Q(source__iexact="hello"))
+        self.assert_query("source:='hello'", Q(source__exact="hello"))
+        self.assert_query('source:="hello world"', Q(source__exact="hello world"))
+        self.assert_query("source:='hello world'", Q(source__exact="hello world"))
+        self.assert_query("source:=hello", Q(source__exact="hello"))
 
     def test_regex(self):
         self.assert_query('source:r"^hello"', Q(source__regex="^hello"))